moving IBaseEntity to common folder
