+++ express-api/src/controllers/admin/agencies/IAgency.ts
@@ -1,5 +1,5 @@
 import { UUID } from 'crypto';
-import { IBaseEntity } from '@/controllers/IBaseEntity';
+import { IBaseEntity } from '@/controllers/common/IBaseEntity';
 
 export interface IAgency extends IBaseEntity {
   id?: UUID;
